# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quizawa",
        "description": "Listing the routes",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "List of users",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Store an user",
                "parameters": [
                    {
                        "name": "pseudo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "stored",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "pseudo or email already exists",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Show one user",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "description": "Delete one user",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful without content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "description": "Update an user by himself or assign a class by admin and creator",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "query",
                        "description": "User authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pseudo",
                        "in": "query",
                        "description": "Pseudo of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "First name of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "Last name of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "admin",
                        "in": "query",
                        "description": "Boolean to define if he's an admin",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "creator",
                        "in": "query",
                        "description": "Boolean to define if he's a creator",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "classroom_id",
                        "in": "query",
                        "description": "Classroom of the user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Login with the user account",
                "parameters": [
                    {
                        "name": "pseudo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "found, authentication success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "parameters are missing",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "not found, authentication fail",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
...
}